Patch from abbotti@mev.co.uk (Ian Abbott):
The patch adds tests for libm in addition to the "boost" tests, as the test for
libcomedi may fail otherwise.

I cheated a bit in the test for libboost_program_options (and also libm) by
checking for the 'main' function, but that seems to be a fairly common trick.
Could check for a function that exists in libm (preferably one that returns
'int') instead of main.  Testing for functions that exist in
libboost_program_options is more difficult!

The patch also adds -lm to the linker flags for comedi_soft_calibrate.