Add place-holder file 'comedi_soft_calibrate/results/README'.
The 'comedi_soft_calibrate/Makefile.am' has a line:

   EXTRA_DIST = results

This is meant to add all the files in the 'results' subdirectory to the
distribution just like 'comedi_calibrate/results'.

Currently, the subdirectory does not exist and Git does cannot store
empty directories (at least not easily) so add a 'README' file as a
place-holder in this subdirectory.

Signed-off-by: Ian Abbott <abbotti@mev.co.uk>