Check driver version before warning in cal_ni_pci_6052e().
The current implementation always prints the warning, regardless of
the comedi driver version.

Also, in cal_ni_pci_6711, there is ambiguity for version 0.7.69, which
will generate a warning message saying that that version is
sufficient.  Either the comparison verision should go down, or the
suggested version should go up, but I'm not sure which.